+++ test/controllers/users_controller_test.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require 'test_helper'
+require "test_helper"
 
-SimpleCov.command_name 'test:controllers'
+SimpleCov.command_name "test:controllers"
 
 # Tests to make sure users can access account settings, and reset password, and
 # that admins can edit and update existing users.
@@ -11,104 +11,102 @@ class UsersControllerTest < ActionController::TestCase
     @user = users(:valid)
   end
 
-  test 'should get index as admin' do
+  test "should get index as admin" do
     login(users(:admin))
     get :index
     assert_not_nil assigns(:users)
     assert_response :success
   end
 
-  test 'should get index for autocomplete' do
+  test "should get index for autocomplete" do
     login(users(:valid))
-    get :index, format: 'json'
+    get :index, format: "json"
     assert_not_nil assigns(:users)
     assert_response :success
   end
 
-  test 'should not get index for non-system admin' do
+  test "should not get index for non-system admin" do
     login(users(:valid))
     get :index
     assert_nil assigns(:users)
-    assert_equal 'You do not have sufficient privileges to access that page.', flash[:alert]
+    assert_equal "You do not have sufficient privileges to access that page.", flash[:alert]
     assert_redirected_to root_path
   end
 
-  test 'should not get index with pagination for non-system admin' do
+  test "should not get index with pagination for non-system admin" do
     login(users(:valid))
-    get :index, format: 'js'
+    get :index, format: "js"
     assert_nil assigns(:users)
-    assert_equal 'You do not have sufficient privileges to access that page.', flash[:alert]
+    assert_equal "You do not have sufficient privileges to access that page.", flash[:alert]
     assert_redirected_to root_path
   end
 
-  test 'should get invite for regular user' do
+  test "should get invite for regular user" do
     login(users(:valid))
-    get :invite, params: { q: 'associated' }
+    get :invite, params: { q: "associated" }
     users_json = JSON.parse(response.body)
-    assert_equal 'associated@example.com', users_json.first['value']
-    assert_equal 'Associated User', users_json.first['name']
+    assert_equal "associated@example.com", users_json.first["value"]
+    assert_equal "Associated User", users_json.first["name"]
     assert_response :success
   end
 
-  test 'should show user as admin' do
+  test "should show user as admin" do
     login(users(:admin))
     get :show, params: { id: @user }
     assert_response :success
   end
 
-  test 'should get edit as admin' do
+  test "should get edit as admin" do
     login(users(:admin))
     get :edit, params: { id: @user }
     assert_response :success
   end
 
-  test 'should not get edit as regular user' do
+  test "should not get edit as regular user" do
     login(users(:valid))
     get :edit, params: { id: @user }
     assert_redirected_to root_path
   end
 
-  test 'should update user as admin' do
+  test "should update user as admin" do
     login(users(:admin))
     patch :update, params: {
       id: @user,
       user: {
-        first_name: 'FirstName',
-        last_name: 'LastName',
-        email: 'valid_updated_email@example.com',
+        full_name: "FirstName LastName",
+        email: "valid_updated_email@example.com",
         system_admin: false
       }
     }
     assert_redirected_to @user
   end
 
-  test 'should not update user with blank name' do
+  test "should not update user with blank full name" do
     login(users(:admin))
     patch :update, params: {
-      id: @user, user: { first_name: '', last_name: '' }
+      id: @user, user: { full_name: "" }
     }
     assert_not_nil assigns(:user)
-    assert_template 'edit'
+    assert_template "edit"
     assert_response :success
   end
 
-  test 'should not update user with invalid id' do
+  test "should not update user with invalid id" do
     login(users(:admin))
     patch :update, params: {
       id: -1,
       user: {
-        first_name: 'FirstName',
-        last_name: 'LastName',
-        email: 'valid_updated_email@example.com',
+        full_name: "FirstName LastName",
+        email: "valid_updated_email@example.com",
         system_admin: false
       }
     }
     assert_redirected_to users_path
   end
 
-  test 'should destroy user as admin' do
+  test "should destroy user as admin" do
     login(users(:admin))
-    assert_difference('User.current.count', -1) do
+    assert_difference("User.current.count", -1) do
       delete :destroy, params: { id: @user }
     end
     assert_redirected_to users_path